ensure pool is never create in recovation callback

+++ arroyo/processing/strategies/run_task_with_multiprocessing.py
@@ -307,14 +307,16 @@ class MultiprocessingPool:
     ) -> None:
         self.__num_processes = num_processes
         self.__initializer = initializer
-        self.__create_pool()
+        self.__pool: Optional[Pool] = None
+        self.maybe_create_pool()
 
-    def __create_pool(self) -> None:
-        self.__pool = Pool(
-            self.__num_processes,
-            initializer=partial(parallel_worker_initializer, self.__initializer),
-            context=multiprocessing.get_context("spawn"),
-        )
+    def maybe_create_pool(self) -> None:
+        if self.__pool is None:
+            self.__pool = Pool(
+                self.__num_processes,
+                initializer=partial(parallel_worker_initializer, self.__initializer),
+                context=multiprocessing.get_context("spawn"),
+            )
 
     @property
     def num_processes(self) -> int:
@@ -325,14 +327,10 @@ class MultiprocessingPool:
         return self.__initializer
 
     def apply_async(self, *args: Any, **kwargs: Any) -> Any:
-        return self.__pool.apply_async(*args, **kwargs)
-
-    def reset(self) -> None:
-        """
-        Close and re-create pool.
-        """
-        self.close()
-        self.__create_pool()
+        if self.__pool:
+            return self.__pool.apply_async(*args, **kwargs)
+        else:
+            raise RuntimeError("No pool available")
 
     def close(self) -> None:
         """
@@ -340,7 +338,9 @@ class MultiprocessingPool:
         Also called from strategy.join() if there are pending futures in order
         ensure state is completely cleaned up.
         """
-        self.__pool.terminate()
+        if self.__pool:
+            self.__pool.terminate()
+            self.__pool = None
 
 
 class RunTaskWithMultiprocessing(
@@ -504,11 +504,12 @@ class RunTaskWithMultiprocessing(
         self.__max_input_block_size = max_input_block_size
         self.__max_output_block_size = max_output_block_size
 
+        self.__pool = pool
+        self.__pool.maybe_create_pool()
+        num_processes = self.__pool.num_processes
+
         self.__shared_memory_manager = SharedMemoryManager()
         self.__shared_memory_manager.start()
-
-        self.__pool = pool
-        num_processes = self.__pool.num_processes
 
         self.__input_blocks = [
             self.__shared_memory_manager.SharedMemory(
@@ -833,7 +834,7 @@ class RunTaskWithMultiprocessing(
         logger.info("Terminating %r...", self.__pool)
 
         logger.info("Shutting down %r...", self.__shared_memory_manager)
-        self.__pool.reset()
+        self.__pool.close()
 
         self.__shared_memory_manager.shutdown()
 
@@ -862,7 +863,7 @@ class RunTaskWithMultiprocessing(
         # XXX: We need to recreate the pool if there are still pending futures, to avoid
         # state from the previous assignment not being properly cleaned up.
         if len(self.__processes):
-            self.__pool.reset()
+            self.__pool.close()
 
         self.__shared_memory_manager.shutdown()
